requirements.txt
pyautogui==0.9.54
pillow==10.0.0
openai==1.3.0
numpy==1.24.3
requests==2.31.0
litellm>=1.0.0